+++ src/ytextractor/video.nim
@@ -1,53 +1,41 @@
 #[
   Created at: 08/09/2021 12:10:05 Monday
-  Modified at: 08/11/2021 03:01:36 PM Wednesday
+  Modified at: 08/14/2021 11:21:29 PM Saturday
 ]#
 
 ##[
   video
   -----
 
-  Parser for `https://www.youtube.com/watch?v=VIDEOID`
+  Parser for "https://www.youtube.com/watch?v=VIDEOID"
 ]##
 
 {.experimental: "codeReordering".}
 
 from std/times import DateTime, Duration, initDuration, now
-from std/json import JsonNode, newJObject, items, hasKey, `{}`, getStr, getInt,
-                     getBool
+from std/json import JsonNode, items, hasKey, `{}`, getStr, getInt, getBool
 from std/strformat import fmt
 from std/strutils import parseInt, multiReplace, find, strip
 
-import ytextractor/base
+import ytextractor/core/core
+from ytextractor/channel import channelId
 
 type
-  YoutubeVideo* = object
+  YoutubeVideo* = object of YoutubeVideoPreview
     ## Youtube video object
-    status*: YoutubeVideoStatus
+    status*: ExtractStatus
     id*: YoutubeVideoId
     title*, description*: string
-    thumbnails*: seq[YoutubeVideoUrl]
-    embed*: YoutubeVideoUrl
+    thumbnails*: seq[UrlAndSize]
+    embed*: UrlAndSize
     publishDate*, uploadDate*: DateTime
     length*: Duration
     familyFriendly*, unlisted*, private*, live*: bool
-    channel*: YoutubeVideoChannel
+    channel*: YoutubeChannelPreview
     views*: int
     category*: YoutubeVideoCategories
     likes*, dislikes*: int
     keywords*: seq[string]
-
-  YoutubeVideoUrl* = object
-    ## Youtube video url object
-    ## Can be the embed url or some image
-    url*: string
-    width*, height*: int
-  YoutubeVideoChannel* = object
-    ## Channel data extracted from video url
-    url*, name*, id*: string
-    subscribers*: int ## This value is not prescise, the Youtube round the value
-    icons*: seq[YoutubeVideoUrl]
-    hiddenSubscribers*: bool
   YoutubeVideoCategories* {.pure.} = enum
     ## Youtube video categories
     Unknown, FilmAndAnimation, AutosAndVehicles, Music, PetsAndAnimals, Sports,
@@ -56,13 +44,6 @@ type
     NonprofitsAndActivism
   YoutubeVideoId* = distinct string
     ## Video Id is a distinct string just for disallow pass any string to parser
-  YoutubeVideoStatus* = object
-    ## Status of parsing update
-    lastUpdate*: DateTime
-    error*: YoutubeVideoError
-  YoutubeVideoError* {.pure.} = enum
-    ## Parsing error
-    None, NotExist, ParseError, InvalidId
 
 proc parseCategory*(str: string): YoutubeVideoCategories =
   ## Parses the category to `YoutubeVideoCategories`
@@ -96,14 +77,6 @@ proc initYoutubeVideo*(id: YoutubeVideoId): YoutubeVideo =
   ## Initialize a new `YoutubeVideo` instance
   YoutubeVideo(id: id)
 
-proc find(nodes: JsonNode; key: string): JsonNode =
-  ## Gets an json array and returns a object
-  ## If not found, returns a empty one
-  result = newJObject()
-  for node in nodes:
-    if node.hasKey key:
-      return node{key}
-
 proc update*(self: var YoutubeVideo; proxy = ""): bool =
   ## Update all `YoutubeVideo` data
   ## Returns `false` on error.
@@ -116,7 +89,7 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
   result = true
 
   if not self.id.valid:
-    self.status.error = YoutubeVideoError.InvalidId
+    self.status.error = ExtractError.InvalidId
     return false
 
   let
@@ -127,7 +100,7 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
     contents = jsonData.ytInitialData{"contents"}
 
   if microformat.isNil:
-    self.status.error = YoutubeVideoError.NotExist
+    self.status.error = ExtractError.NotExist
     return false
 
   self.status.lastUpdate = now()
@@ -137,7 +110,7 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
     self.description = microformat{"description", "simpleText"}.getStr
     block thumbnail:
       for thumb in videoDetails{"thumbnail", "thumbnails"}:
-        self.thumbnails.add YoutubeVideoUrl(
+        self.thumbnails.add UrlAndSize(
           url: thumb{"url"}.getStr,
           width: thumb{"width"}.getInt,
           height: thumb{"height"}.getInt,
@@ -156,22 +129,14 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
     self.unlisted = microformat{"isUnlisted"}.getBool
     self.category = microformat{"category"}.getStr.parseCategory
     block channel:
-      self.channel.url = microformat{"ownerProfileUrl"}.getStr
-      self.channel.id = microformat{"externalChannelId"}.getStr
+      self.channel.id = microformat{"ownerProfileUrl"}.getStr.channelId
       self.channel.name = microformat{"ownerChannelName"}.getStr
 
       let subs = contents{"twoColumnWatchNextResults", "results", "results",
                           "contents"}.
-        find("videoSecondaryInfoRenderer"){"owner", "videoOwnerRenderer",
+        findInJson("videoSecondaryInfoRenderer"){"owner", "videoOwnerRenderer",
             "subscriberCountText", "simpleText"}.
-          getStr.multiReplace({
-              "K": "000",
-              "M": "000000",
-              " million": "000000",
-              "B": "000000000",
-              " subscribers": "",
-              ".": ""
-            })
+          getStr.parseSubs
 
       if subs.len == 0:
         self.channel.hiddenSubscribers = true
@@ -181,9 +146,9 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
       block channelIcons:
         for icon in contents{"twoColumnWatchNextResults", "results", "results",
                             "contents"}.
-          find("videoSecondaryInfoRenderer"){"owner", "videoOwnerRenderer",
+          findInJson("videoSecondaryInfoRenderer"){"owner", "videoOwnerRenderer",
               "thumbnail", "thumbnails"}:
-          self.channel.icons.add YoutubeVideoUrl(
+          self.channel.icons.add UrlAndSize(
             url: icon{"url"}.getStr,
             width: icon{"width"}.getInt,
             height: icon{"height"}.getInt,
@@ -192,7 +157,7 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
     block likes:
       let data = contents{"twoColumnWatchNextResults", "results", "results",
                           "contents"}.
-        find("videoPrimaryInfoRenderer"){"videoActions", "menuRenderer",
+        findInJson("videoPrimaryInfoRenderer"){"videoActions", "menuRenderer",
             "topLevelButtons"}
       proc get(data: JsonNode; i: int): int {.inline.} =
         data{i}{"toggleButtonRenderer", "defaultText", "accessibility",
@@ -215,9 +180,9 @@ proc update*(self: var YoutubeVideo; proxy = ""): bool =
     self.private = videoDetails{"isPrivate"}.getBool
     self.live = videoDetails{"isLiveContent"}.getBool
 
-    self.status.error = YoutubeVideoError.None
+    self.status.error = ExtractError.None
   except:
-    self.status.error = YoutubeVideoError.ParseError
+    self.status.error = ExtractError.ParseError
     # doAssert false, getCurrentExceptionMsg()
     return false
 
@@ -272,7 +237,6 @@ proc extractVideo*(url: string; proxy = ""): YoutubeVideo =
   ##
   ## .. code-block:: nim
   ##   var vid = initYoutubeVideo("jjEQ-yKVPMg".videoId)
-  ##   test: "echo Hello world"
   ##   discard vid.update():
   ## **Example:**
   ##